Base report class and application classes
The commit introduces two base report classes: one to create the a custom formatter, the latter to wrap the formatter rapidly in an application.

The approach has been tested on the json formatter.

An initial scaffolding for the default text formatter has been provided, but implementation has not been done (prints only "not yet implemented")

The approach implements also:

 - command line provided formatter parameters
 - support for reading from stdin
 - support for dump into stdout and stderr

Signed-off-by: Matteo Ragni <matteo.ragni@antemotion.com>